requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements-dev.txt requirements-dev.in
#
anyio==3.3.1
    # via httpcore
attrs==21.2.0
    # via pytest
black==21.9b0
    # via -r requirements-dev.in
certifi==2021.5.30
    # via
    #   httpx
    #   requests
charset-normalizer==2.0.6
    # via
    #   httpx
    #   requests
click==8.0.1
    # via
    #   black
    #   pip-tools
flake8==4.0.1
    # via -r requirements-dev.in
h11==0.12.0
    # via httpcore
httpcore==0.13.7
    # via httpx
httpx==0.20.0
    # via -r requirements-dev.in
idna==3.2
    # via
    #   anyio
    #   requests
    #   rfc3986
iniconfig==1.1.1
    # via pytest
isort==5.9.3
    # via -r requirements-dev.in
mccabe==0.6.1
    # via flake8
mypy==0.910
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.0
    # via pytest
pathspec==0.9.0
    # via black
pep517==0.11.0
    # via pip-tools
pip-tools==6.4.0
    # via -r requirements-dev.in
platformdirs==2.3.0
    # via black
pluggy==1.0.0
    # via pytest
py==1.10.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via -r requirements-dev.in
regex==2021.8.28
    # via black
requests==2.26.0
    # via -r requirements-dev.in
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
toml==0.10.2
    # via
    #   mypy
    #   pytest
tomli==1.2.1
    # via
    #   black
    #   pep517
typing-extensions==3.10.0.2
    # via
    #   black
    #   mypy
urllib3==1.26.6
    # via requests
wheel==0.37.0
    # via pip-tools
sqlalchemy==1.4.23
cryptography==3.1.1  

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
uvicorn==0.15.0
fastapi_mail==1.0.1
mysql-connector-python==8.0.28
MySQL-python==1.2.5